Use Delta format

+++ src/kexp/aws_layers/spark_handler/core.py
@@ -16,6 +16,7 @@
 # ------------------------------------------------
 
 from pyspark.sql import SparkSession
+from pyspark.sql import SQLContext
 from delta import *
 
 from os import listdir
@@ -33,19 +34,24 @@ class SparkCore:
     spark = None
     sql_context = None
 
-    def __init__(self, app_name, root_location):
+    def __init__(self, app_name, lake_location):
         self.app_name = app_name
-        self.bronze_location = f"{root_location}/bronze"
-        self.silver_location = f"{root_location}/silver"
-        self.gold_location = f"{root_location}/gold"
+        self.bronze_location = f"{lake_location}/bronze"
+        self.silver_location = f"{lake_location}/silver"
+        self.gold_location = f"{lake_location}/gold"
 
         builder = SparkSession.builder.appName(app_name) \
             .config("spark.sql.extensions", "io.delta.sql.DeltaSparkSessionExtension") \
             .config("spark.sql.catalog.spark_catalog", "org.apache.spark.sql.delta.catalog.DeltaCatalog")
 
         self.spark = configure_spark_with_delta_pip(builder).getOrCreate()
+        self.sql_context = SQLContext(self.spark.sparkContext)
 
-    def process_raw_to_bronze(self, raw_data_folder, file_match):
+    def process_raw_to_bronze(self, raw_data_folder, file_match, replace=True):
+        """
+        Process the raw data and if replace is true, then delete the existing object.
+        """
+
         result = {"raw": [], "bronze": []}
 
         spark = SparkSession.builder.getOrCreate()
@@ -58,13 +64,17 @@ class SparkCore:
             df = spark.read.load(join(raw_data_folder, file),
                                  format=file_type, inferSchema="true", header="true")
 
-            try:
-                shutil.rmtree(join(self.bronze_location, file_name))
-            except OSError as e:
-                print("Warning: %s : %s" % (join(self.bronze_location, file_name), e.strerror))
+            if replace:
+                try:
+                    shutil.rmtree(join(self.bronze_location, file_name))
+                except OSError as e:
+                    print("Warning: %s : %s" % (join(self.bronze_location, file_name), e.strerror))
 
             df.write.format("delta").save(join(self.bronze_location, file_name))
 
             result["bronze"].append(join(self.bronze_location, file_name))
 
         return result
+
+    def sql(self, sql_statement):
+        return self.sql_context.sql(sql_statement)
